backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
alembic
python-jose[cryptography]
passlib[argon2]
pydantic[email]
python-multipart